# docs/swagger.yaml
basePath: /cs/v1
definitions:
...
  course-service_internal_domain.Course:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      ownerUserId:
        type: integer
      stats:
        $ref: '#/definitions/course-service_internal_domain.CourseStats'
      status:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  course-service_internal_domain.CourseSectionItem:
    properties:
      id:
        type: integer
      is_published:
        type: boolean
      item_type:
        type: string
      section_id:
        type: integer
      sort_order:
        type: integer
      title:
        type: string
    type: object
  course-service_internal_domain.CourseSectionWithItems:
    properties:
      course_id:
        type: integer
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/course-service_internal_domain.CourseSectionItem'
        type: array
      parent_id:
        type: integer
      sort_order:
        type: integer
      title:
        type: string
    type: object
  course-service_internal_domain.CourseStats:
    properties:
      totalStudents:
        type: integer
    type: object
  course-service_internal_domain.CourseWithItems:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      ownerUserId:
        type: integer
      sections:
        items:
          $ref: '#/definitions/course-service_internal_domain.CourseSectionWithItems'
        type: array
      stats:
        $ref: '#/definitions/course-service_internal_domain.CourseStats'
      status:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  course-service_internal_transport_http_v1_request.CreateCourse:
    properties:
      description:
        type: string
      ownerUserId:
        type: integer
      title:
        type: string
    type: object
...
  course-service_internal_transport_http_v1_request.DeleteCourseListeners:
    properties:
      groupIds:
        items:
          type: integer
        type: array
      userIds:
        items:
          type: integer
        type: array
    type: object
...
  course-service_internal_transport_http_v1_request.SetCourseListener:
    properties:
      groupIds:
        items:
          type: integer
        type: array
      userIds:
        items:
          type: integer
        type: array
    type: object
...
  course-service_internal_transport_http_v1_request.UpdateCourse:
    properties:
      description:
        type: string
      ownerUserId:
        type: integer
      status:
        type: integer
      title:
        type: string
    type: object
...
  internal_transport_http_v1.ErrorResponse:
    properties:
      message:
        example: bad request
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for courses, sections, assignments, question banks and tests.
    Tests are also available under `/api` (same handlers as `/cs/v1`).
  title: Course Service API
  version: "1.0"
paths:
...
  /courses:
    post:
      consumes:
      - application/json
      parameters:
      - description: Course
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.CreateCourse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course-service_internal_domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Create course
      tags:
      - courses
  /courses/{course_id}/listeners:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - description: Listeners
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.DeleteCourseListeners'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Remove course listeners
      tags:
      - courses
    get:
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              format: int64
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List course listeners
      tags:
      - courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - description: Listeners
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.SetCourseListener'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Add course listeners
      tags:
      - courses
  /courses/{course_id}/with-all-items:
    get:
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.CourseWithItems'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get course with sections and items
      tags:
      - courses
  /courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Delete course
      tags:
      - courses
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get course by ID
      tags:
      - courses
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Course
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateCourse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Update course
      tags:
      - courses
...
schemes:
- http
swagger: "2.0"